+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Fixed
 - Draw extension line correctly if stop object is not next row.
 - Play tempo/dynamics correctly without extension line.
+- DocumentBuidler function argument error string.
 
 ## [6.0.0-pre.5] - 2025-12-14
 ### Changed